Move homepage site summary items into their respective apps
