+++ wagtail/wagtailimages/templates/wagtailimages/homepage/site_summary_images.html
